Improve minmake scripts
Introduce new rule for install rule_inst, taking two arguments,
the file to install and the target directory. (Installing a file
under a different name is currently not supported)
The rule is called via the new echo-cmd-arg2 call, that now
takes 2 extra arguments. And there is a normal and quiet echo
version, the quiet one simply says INSTALL <file>

Next we replace all the appropriate install calls in our
templates by calls to rule_inst

Only install a library or binary if its -Instdir magic variable
is not empty. That way we can build test binaries that are not
not installed by setting their instdir to empty.

That all makes the output of make install much nicer.

We need to export our MM-default-bindir to have the desired effect.

During make install and make build, and only in non-verbose mode,
we now echo the current directory when that changes. Otherwise,
when "make build" has almost nothing to do it would not report
progress and if it then fails, the user has no clue where.

The recursive make commands itself are now hidden in verbose mode,
because make's directory changing messages are already in effect.