minmake improvement
Create a makefile that automatically calls the toplevel make
with the appropriate sub directory.

The file is used by creating a symlink to it that is
named GNUmakefile, i.e.::

   <topdir>/scripts>
   <topdir>/sub/dir/GNUmakefile -> ../../scripts/call-subdir-make

When make is now called in <topdir>/sub/dir, it first finds
GNUmakefile and will use it, ignoring the normal Makefile.

This file will now determine the project toplevel directory by
resolving the symlink GNUmakefile in the current directory,
and stepping one directory level up. It will also determine the
current subdirectory it is called from by resolving . and
subtracting the project toplevel directory from that path.
Finall it will call the toplevel make with the original arguments
plus D=<subdir>

There is a small trick involved: We don't know what targets are
possible, so we must pretend that any target is possible, that is what
the %: target ist for. But if several targets are given, make would call
the recipe for each target, which is not what we want, so the default
recipe does nothing, but uses a prerequisite that performs the actual
action and will only called once.