Minmake recursion cleanup and subdir fix
Recursion was done a bit different, in the top-level and in the
subdirectories, leading to much duplicated code.

The main difference has been the subdirectory selection, which either
was "$@" in case dir is empty, or "$(dir)/$@" otherwise, so move
consolidate that into the _MM_subdir variable using an $(if )
expression.

Remove the _all and all pseudo targets in favour of _build and build

Making a subdirectory only got a small fix: We use the same trick as
in call-subdir-make, to only call a single make if several targets
are given on the commandline. So no you can say:

  make D=sub/dir clean build

to rebuild the given subdirectory.